backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "temple-cat-voice-backend"
version = "0.0.0"
description = "Voice AI backend"
readme = "../interview-project-details.md"
dependencies = [
    "fastapi>=0.115.4",
    "pydantic>=2.9.2",
    "uvicorn>=0.32.0",
    "httpx>=0.27.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pipecat-ai[daily,deepgram,openai,cartesia,silero]>=0.0.50",
    "aiohttp>=3.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
